Refactor the `tbtc/storage.go` file
